refactor(types): checkbox
